UICIRC-906 - create edit sections
