+++ src/settings/PatronNotices/components/EditSections/PatronNoticeAboutSection/index.js
@@ -1,0 +1,1 @@
+export { default } from './PatronNoticeAboutSection';
