+++ src/settings/PatronNotices/components/EditSections/PatronNoticeEmailSection/index.js
@@ -1,0 +1,1 @@
+export { default } from './PatronNoticeEmailSection';
