+++ src/settings/PatronNotices/components/EditSections/index.js
@@ -1,0 +1,2 @@
+export { default as PatronNoticeAboutSection } from './PatronNoticeAboutSection';
+export { default as PatronNoticeEmailSection } from './PatronNoticeEmailSection';
